--- pages/components/layout.module.css
@@ -1,15 +1,0 @@
-.container {
-    max-width: 36rem;
-    padding: 0 1rem;
-    margin: 3rem auto 6rem;
-}
-
-.header {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.backToHome {
-    margin: 3rem 0 0;
-}
