+++ pages/index.js
@@ -1,9 +1,9 @@
 import Head from 'next/head';
-import Layout, {siteTitle} from './components/layout';
+import Layout, {siteTitle} from '../components/layout';
 import utilStyles from '../styles/utils.module.css';
 import { getSortedPostsData } from '../lib/posts';
 import Link from 'next/link';
-import Date from './components/date';
+import Date from '../components/date';
 
 export async function getStaticProps() {
   const allPostsData = getSortedPostsData();
